Make providing `--kms` a bit more forgiving
A KMS URI can now be provided using just the scheme, without
a colon. E.g. `--kms tpmkms` is now a valid value.

Before this change, providing `--kms tpmkms` would result in an
error: `Error: error parsing tpmkms: scheme is missing`. It isn't
immediately clear that it's the colon that's missing, so decided
to make the KMS URI processing a bit smarter.