Add (experimental) support for creating new (attested) keys in `attest`
To make it possible to use the `attest` command similar to how it's
used with a YubiKey, for TPMs we need to create a new attested key
at the time of creating an attestation. This is because the data
that is being attested when attesting the key is part of the input
for the TPM when creating the key.

It's still possible to do a normal `attest` flow. That requires
creating an AK and key (with optional `qualifying-data` set) to
be created first. Using the `--new` flag, a new key with `name`
is created. This option is not yet documented in an example.