Upgrade goreleaser-cross to v1.24
This commit upgrades goreleaser-cross to the latest version. This
version includes a new version of goreleaser that deprecates some
parameters, this commit upgrades those parameters to use the new
version.

This should fix an issue with the CI when go builds the go toolchain
because the go.mod was using a newer version.